pyproject.toml
[project]
name = "gtmlens"
version = "0.1.0"
description = "Causal targeting engine for GTM funnels"
requires-python = ">=3.12"

dependencies = [
    # Data layer
    "duckdb>=1.5",
    "pandas>=2.0",
    "numpy>=1.26",
    "faker>=20.0",
    "scikit-learn>=1.3",
    "scipy>=1.11",
    # API layer
    "fastapi[standard]>=0.100",
    "uvicorn[standard]>=0.24",
    "python-multipart>=0.0.9",
    "pydantic[email]>=2.0",
    # Auth
    "bcrypt>=4.0",
    "python-jose[cryptography]>=3.3",
    # Utilities
    "python-dotenv>=1.0",
    "econml>=0.16.0",
    "statsmodels>=0.14.6",
    "anthropic>=0.95.0",
    "resend>=2.28.0",
]

[dependency-groups]
dev = [
    "pytest>=9.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["api", "core", "data", "ingestion"]

[tool.pytest.ini_options]
testpaths = ["tests"]
